+++ CHANGELOG.md
@@ -3,4 +3,5 @@
 - [CHANGE] Rename label in `bind_incoming_requests_total` from `name` to `opcode`
 - [CHANGE] Rename flag `-bind.statsuri` to `-bind.stats-url`
 - [FEATURE] Add support for BIND statistics v3
+- [FEATURE] Automatically detect BIND statistics version and use correct client
 - [FEATURE] Provide option to control exported statistics with `-bind.stats-groups`